+++ client/mapview.c
@@ -1478,7 +1478,7 @@ void scrollbar_jump_callback(Widget w, XtPointer client_data,
 void scrollbar_scroll_callback(Widget w, XtPointer client_data,
 			     XtPointer position_val)
 {
-  int position=(int)position_val;
+  size_t position=(size_t)position_val;
 
 
   if(get_client_state()!=CLIENT_GAME_RUNNING_STATE)
