+++ client/ratesdlg.c
@@ -361,7 +361,7 @@ void rates_scroll_jump_callback(Widget w, XtPointer client_data,
 void rates_scroll_scroll_callback(Widget w, XtPointer client_data,
 				  XtPointer position_val)
 {
-  int pos=(int)position_val;
+  size_t pos=(size_t)position_val;
   int val;
   
   if(w==rates_tax_scroll) {
